norme progressing
